requirements.txt
openai
python-dotenv
python-docx
pyPDF2
pdfplumber
gradio
pycord
discord
azure-core
azure-ai-formrecognizer
